[back-end] Resolved #9: Refactoring. ICategoryRepository

+++ back-end/Infrastructure/Repository/CategoryRepository.cs
@@ -11,7 +11,6 @@ namespace Infrastructure.Repository
 {
     public class CategoryRepository : ICategoryRepository
     {
-
         private readonly DataContext _context;
 
         public CategoryRepository(DataContext context)
@@ -19,23 +18,29 @@ namespace Infrastructure.Repository
             _context = context;
         }
 
-        public async Task<bool> CheckIsEverything(string code)
-        {
-            var item = await GetEverythingCategory();
 
-            return item?.Code == code;
+
+        public Task<int> CountAsync()
+        {
+            return _context.Categories.CountAsync();
         }
+
 
         public async Task<bool> DeleteAsync(Category category)
         {
+            if (category.Id == null)
+                throw new Exception($"Can't delete new category");
+
             var dbCategory = await _context.Categories.FirstOrDefaultAsync(x => x.Id == category.Id);
+            if (category.Id == null)
+                throw new Exception($"Can't find category with id: {category.Id}");
 
             _context.Categories.Remove(dbCategory);
             var rows = await _context.SaveChangesAsync();
 
-
             return rows == 1;
         }
+
 
         public Task<Category[]> GetAsync()
         {
@@ -44,25 +49,22 @@ namespace Infrastructure.Repository
                            .Select(x => DataConverter.ToCategory(x))
                            .ToArrayAsync();
         }
-        
-        public Task<Category> GetAsync(string code)
-        {
-            return GetCategory(x => x.Code == code);
-        }
 
         public Task<Category> GetAsync(int id)
         {
             return GetCategory(x => x.Id == id);
         }
 
-        public Task<Category> GetEverythingCategory()
+        public Task<Category> GetAsync(string code)
         {
-            return _context.CategoriesWithTotalProjects
-                           .AsNoTracking()
-                           .Where(x=>x.IsEverything)
-                           .Select(x => DataConverter.ToCategory(x))
-                           .FirstOrDefaultAsync();
+            return GetCategory(x => x.Code == code);
         }
+
+        public Task<Category> GetTechnicalAsync()
+        {
+            return GetCategory(x => x.IsEverything);
+        }
+
 
         public Task<Category> SaveAsync(Category item)
         {
@@ -72,7 +74,8 @@ namespace Infrastructure.Repository
             return UpdateAsync(item);
         }
 
-        public async Task<Category> CreateAsync(Category category)
+
+        private async Task<Category> CreateAsync(Category category)
         {
             //TODO: should create dedicated SaveNew method
             if (string.IsNullOrEmpty(category.Code))
@@ -103,7 +106,7 @@ namespace Infrastructure.Repository
             return DataConverter.ToCategory(dbItem);
         }
 
-        public async Task<Category> UpdateAsync(Category category)
+        private async Task<Category> UpdateAsync(Category category)
         {
             if (category.Id != null)
                 throw new Exception("Inconsistency");
@@ -124,7 +127,6 @@ namespace Infrastructure.Repository
             return DataConverter.ToCategory(dbItem);
         }
 
-
         private async Task<Category> GetCategory(Expression<Func<DataModel.CategoryWithTotalProjects, bool>> predicate)
         {
             var result = await _context.CategoriesWithTotalProjects
@@ -133,13 +135,8 @@ namespace Infrastructure.Repository
                .Select(x => DataConverter.ToCategory(x))
                .FirstOrDefaultAsync();
 
-            if (result == null)
-                throw new Exception("Not found");
-
             return result;
         }
-
-
        
     }
 }
